Crate sr_primitives

Runtime Modules shared primitive types.

Modules

biguint
Re-export big_uint stuff. Infinite precision unsigned integer for substrate runtime.

curve
Provides some utilities to define a piecewise linear function.

generic
Generic implementations of Extrinsic/Header/Block.

helpers_128bit
Re-export 128 bit helpers. Some helper functions to work with 128bit numbers. Note that the functionality provided here is only sensible to use with 128bit numbers because for smaller sizes, you can always rely on assumptions of a bigger type (u128) being available, or simply create a per-thing and use the multiplication implementation provided there.

key_types
Re-export this since it's part of the API of this crate. Known key types; this also functions as a global registry of key types for projects wishing to avoid collisions with each other.

offchain
A collection of higher lever helpers for offchain workers.

testing
Testing utilities.

traits
Primitives for the runtime modules.

transaction_validity
Transaction validity interface.

Macros

__impl_outer_config_types
Helper macro for impl_outer_config

assert_eq_error_rate
Checks that $x is equal to $y with an error rate of $error.

count
Calls a given macro a number of times with a set of fixed params and an incrementing numeral. e.g.

create_runtime_str
Create a const RuntimeString.

impl_opaque_keys
Implement OpaqueKeys for a described struct.

impl_outer_config
Implement the output "meta" module configuration struct, which is basically: pub struct GenesisConfig { rust_module_one: Option, ... }

Structs

AccountId32
Re-export this since it's part of the API of this crate. An opaque 32-byte cryptographic identifier.

AnySignature
Signature verify that can work with any known signature types..

Digest
Generic header digest.

DispatchError
Reason why a dispatch call failed

Fixed64
Re-export top-level arithmetic stuff. An unsigned fixed point number. Can hold any value in the range [-9_223_372_036, 9_223_372_036] with fixed point accuracy of one billion.

KeyTypeId
Re-export this since it's part of the API of this crate. An identifier for a type of cryptographic key.

ModuleId
A module identifier. These are per module and should be stored in a registry somewhere.

OpaqueExtrinsic
Simple blob to hold an extrinsic without committing to its format and ensure it is serialized correctly.

Perbill
Re-export top-level arithmetic stuff. A fixed point representation of a number between in the range [0, 1].

Percent
Re-export top-level arithmetic stuff. A fixed point representation of a number between in the range [0, 1].

Permill
Re-export top-level arithmetic stuff. A fixed point representation of a number between in the range [0, 1].

Perquintill
Re-export top-level arithmetic stuff. A fixed point representation of a number between in the range [0, 1].

Rational128
Re-export top-level arithmetic stuff. A wrapper for any rational number with a 128 bit numerator and denominator.

Enums

ApplyError
Reason why an extrinsic couldn't be applied (i.e. invalid extrinsic).

DigestItem
Digest item that is able to encode/decode 'system' digest items and provide opaque access to other items.

MultiSignature
Signature verify that can work with any known signature types..

MultiSigner
Public key for any known crypto algorithm.

Traits

BoundToRuntimeAppPublic
Something that bound to a fixed RuntimeAppPublic.

BuildModuleGenesisStorage
Something that can build the genesis storage of a module.

BuildStorage
Complex storage builder stuff.

CryptoType
Re-export this since it's part of the API of this crate. Type which has a particular kind of crypto associated with it.

Deserialize
A data structure that can be deserialized from any data format supported by Serde.

DeserializeOwned
A data structure that can be deserialized without borrowing any data from the deserializer.

RuntimeAppPublic
A runtime interface for an application's public key.

Serialize
A data structure that can be serialized into any data format supported by Serde.

TypeId
Re-export this since it's part of the API of this crate. Provide a simple 4 byte identifier for a type.

Functions

print
Print something that implements Printable from the runtime.

verify_encoded_lazy
Verify a signature on an encoded value in a lazy manner. This can be an optimization if the signature scheme has an "unsigned" escape hash.

Type Definitions

ApplyOutcome
The outcome of applying a transaction.

ApplyResult
Result from attempt to apply an extrinsic.

ChildrenStorageOverlay
A set of key value pairs for children storage;

ConsensusEngineId
Consensus engine unique ID.

Justification
An abstraction over justification for a block's validity under a consensus algorithm.

RuntimeString
A String that is a &'static str on no_std and a Cow<'static, str> on std.

StorageOverlay
A set of key value pairs for storage.

Derive Macros

Deserialize

RuntimeDebug
Re-export RuntimeDebug, to avoid dependency clutter.

Serialize