Trait sr_primitives::RuntimeAppPublic

pub trait RuntimeAppPublic {
    type Signature: MaybeDebugHash + PartialEq<Self::Signature> + Eq + Clone + Codec;

    const ID: KeyTypeId;

    fn all() -> Vec<Self>;

    fn generate_pair(seed: Option<&str>) -> Self;

    fn sign<M>(&self, msg: &M) -> Option<Self::Signature>
    where
        M: AsRef<[u8]>;

    fn verify<M>(&self, msg: &M, signature: &Self::Signature) -> bool
    where
        M: AsRef<[u8]>;
}

A runtime interface for an application's public key.

Associated Types

type Signature: MaybeDebugHash + PartialEq<Self::Signature> + Eq + Clone + Codec

The signature that will be generated when signing with the corresponding private key.

Loading content...
Associated Constants

const ID: KeyTypeId

An identifier for this application-specific key type.

Loading content...
Required methods

fn all() -> Vec<Self>

Returns all public keys for this application in the keystore.

fn generate_pair(seed: Option<&str>) -> Self

Generate a public/private pair and store it in the keystore.

Returns the generated public key.

fn sign<M>(&self, msg: &M) -> Option<Self::Signature> where
M: AsRef<[u8]>,

Sign the given message with the corresponding private key of this public key.

The private key will be requested from the keystore.

Returns the signature or None if the private key could not be found or some other error occurred.

fn verify<M>(&self, msg: &M, signature: &Self::Signature) -> bool where
M: AsRef<[u8]>,

Verify that the given signature matches the given message using this public key.

Loading content...