Trait sr_primitives::traits::Applyable

pub trait Applyable: Sized + Send + Sync {
    type AccountId: Member + MaybeDisplay;
    type Call;
    fn sender(&self) -> Option<&Self::AccountId>;

    fn validate<V: ValidateUnsigned<Call = Self::Call>>(
        &self, 
        info: DispatchInfo, 
        len: usize
    ) -> TransactionValidity;

    fn dispatch(
        self, 
        info: DispatchInfo, 
        len: usize
    ) -> Result<DispatchResult, DispatchError>;
}

An "executable" piece of information, used by the standard Substrate Executive in order to enact a piece of extrinsic information by marshalling and dispatching to a named function call.

Also provides information on to whom this information is attributable and an index that allows each piece of attributable information to be disambiguated.

Associated Types

type AccountId: Member + MaybeDisplay

Id of the account that is responsible for this piece of information (sender).

type Call

Type by which we can dispatch. Restricts the UnsignedValidator type.

Loading content...
Required methods

fn sender(&self) -> Option<&Self::AccountId>

Returns a reference to the sender if any.

fn validate<V: ValidateUnsigned<Call = Self::Call>>(
&self,
info: DispatchInfo,
len: usize
) -> TransactionValidity

Checks to see if this is a valid transaction. It returns information on it if so.

fn dispatch(
self,
info: DispatchInfo,
len: usize
) -> Result<DispatchResult, DispatchError>

Executes all necessary logic needed prior to dispatch and deconstructs into function call, index and sender.

Loading content...