Trait sr_primitives::traits::As

pub trait As<T> {
    fn as_(self) -> T;

    fn sa(_: T) -> Self;
}

Simple trait similar to Into, except that it can be used to convert numerics between each other.

Required methods

fn as_(self) -> T

Convert forward (ala Into::into).

fn sa(_: T) -> Self

Convert backward (ala From::from).

Loading content...