Trait sr_primitives::traits::Block

pub trait Block: Clone + Send + Sync + Codec + Eq + MaybeSerializeDebugButNotDeserialize + 'static {
    type Extrinsic: Member + Codec + Extrinsic + MaybeSerialize;
    type Header: Header<Hash = Self::Hash>;
    type Hash: Member + MaybeSerializeDebug + Hash + Copy + MaybeDisplay + Default + SimpleBitOps + Codec + AsRef<[u8]> + AsMut<[u8]>;
    fn header(&self) -> &Self::Header;

    fn extrinsics(&self) -> &[Self::Extrinsic];

    fn deconstruct(self) -> (Self::Header, Vec<Self::Extrinsic>);

    fn new(header: Self::Header, extrinsics: Vec<Self::Extrinsic>) -> Self;

    fn hash(&self) -> Self::Hash { ... }
}

Something which fulfills the abstract idea of a Substrate block. It has types for an Extrinsic piece of information as well as a Header.

You can get an iterator over each of the extrinsics and retrieve the header.

Associated Types

type Extrinsic: Member + Codec + Extrinsic + MaybeSerialize

Type of extrinsics.

type Header: Header<Hash = Self::Hash>

Header type.

type Hash: Member + MaybeSerializeDebug + Hash + Copy + MaybeDisplay + Default + SimpleBitOps + Codec + AsRef<[u8]> + AsMut<[u8]>

Block hash type.

Loading content...
Required methods

fn header(&self) -> &Self::Header

Returns a reference to the header.

fn extrinsics(&self) -> &[Self::Extrinsic]

Returns a reference to the list of extrinsics.

fn deconstruct(self) -> (Self::Header, Vec<Self::Extrinsic>)

Split the block into header and list of extrinsics.

fn new(header: Self::Header, extrinsics: Vec<Self::Extrinsic>) -> Self

Creates new block from header and extrinsics.

Loading content...
Provided methods

fn hash(&self) -> Self::Hash

Returns the hash of the block.

Loading content...