Trait sr_primitives::traits::Hash

pub trait Hash: 'static + MaybeSerializeDebug + Clone + Eq + PartialEq {
    type Output: Member + MaybeSerializeDebug + Hash + AsRef<[u8]> + AsMut<[u8]> + Copy + Default + Encode + Decode;
    type Hasher: Hasher<Out = Self::Output>;
    fn hash(s: &[u8]) -> Self::Output;

    fn enumerated_trie_root(items: &[&[u8]]) -> Self::Output;

    fn ordered_trie_root<I: IntoIterator<Item = A> + Iterator<Item = A>, A: AsRef<[u8]>>(
        input: I
    ) -> Self::Output;

    fn trie_root<I: IntoIterator<Item = (A, B)>, A: AsRef<[u8]> + Ord, B: AsRef<[u8]>>(
        input: I
    ) -> Self::Output;

    fn storage_root() -> Self::Output;

    fn storage_changes_root(parent_hash: Self::Output) -> Option<Self::Output>;

    fn hash_of<S: Codec>(s: &S) -> Self::Output { ... }
}

Abstraction around hashing

Associated Types

type Output: Member + MaybeSerializeDebug + Hash + AsRef<[u8]> + AsMut<[u8]> + Copy + Default + Encode + Decode

The hash type produced.

type Hasher: Hasher<Out = Self::Output>

The associated hash_db Hasher type.

Loading content...
Required methods

fn hash(s: &[u8]) -> Self::Output

Produce the hash of some byte-slice.

fn enumerated_trie_root(items: &[&[u8]]) -> Self::Output

Produce the trie-db root of a mapping from indices to byte slices.

fn ordered_trie_root<I: IntoIterator<Item = A> + Iterator<Item = A>, A: AsRef<[u8]>>(
input: I
) -> Self::Output

Iterator-based version of enumerated_trie_root.

fn trie_root<I: IntoIterator<Item = (A, B)>, A: AsRef<[u8]> + Ord, B: AsRef<[u8]>>(
input: I
) -> Self::Output

The Patricia tree root of the given mapping as an iterator.

fn storage_root() -> Self::Output

Acquire the global storage root.

fn storage_changes_root(parent_hash: Self::Output) -> Option<Self::Output>

Acquire the global storage changes root.

Loading content...
Provided methods

fn hash_of<S: Codec>(s: &S) -> Self::Output

Produce the hash of some codec-encodable value.

Loading content...