Trait sr_primitives::traits::ModuleDispatchError

pub trait ModuleDispatchError {
    fn as_u8(&self) -> u8;

    fn as_str(&self) -> &'static str;
}

An error that is returned by a dispatchable function of a module.

Required methods

fn as_u8(&self) -> u8

Convert this error to an u8.

The u8 corresponds to the index of the variant in the error enum.

fn as_str(&self) -> &'static str

Convert the error to a &'static str.

Loading content...