Trait sr_primitives::traits::OffchainWorker

pub trait OffchainWorker<BlockNumber> {
    fn generate_extrinsics(_n: BlockNumber) { ... }
}

Off-chain computation trait.

Implementing this trait on a module allows you to perform a long-running tasks that make validators generate extrinsics (either transactions or inherents) with results of those long-running computations.

NOTE: This function runs off-chain, so it can access the block state, but cannot preform any alterations.

Provided methods

fn generate_extrinsics(_n: BlockNumber)

This function is being called on every block.

Implement this and use special externs to generate transactions or inherents. Any state alterations are lost and are not persisted.

Loading content...