Trait sr_primitives::traits::OnInitialize

pub trait OnInitialize<BlockNumber> {
    fn on_initialize(_n: BlockNumber) { ... }
}

The block initialization trait. Implementing this lets you express what should happen for your module when the block is beginning (right before the first extrinsic is executed).

Provided methods

fn on_initialize(_n: BlockNumber)

The block is being initialized. Implement to have something happen.

Loading content...