Trait sr_primitives::traits::SignedExtension

pub trait SignedExtension: Codec + MaybeDebug + Sync + Send + Clone + Eq + PartialEq {
    type AccountId;
    type Call;
    type AdditionalSigned: Encode;
    type Pre: Default;
    fn additional_signed(&self) -> Result<Self::AdditionalSigned, &'static str>;

    fn validate(
        &self, 
        _who: &Self::AccountId, 
        _call: &Self::Call, 
        _info: DispatchInfo, 
        _len: usize
    ) -> Result<ValidTransaction, DispatchError> { ... }

    fn pre_dispatch(
        self, 
        who: &Self::AccountId, 
        call: &Self::Call, 
        info: DispatchInfo, 
        len: usize
    ) -> Result<Self::Pre, DispatchError> { ... }

    fn validate_unsigned(
        _call: &Self::Call, 
        _info: DispatchInfo, 
        _len: usize
    ) -> Result<ValidTransaction, DispatchError> { ... }

    fn pre_dispatch_unsigned(
        call: &Self::Call, 
        info: DispatchInfo, 
        len: usize
    ) -> Result<Self::Pre, DispatchError> { ... }

    fn post_dispatch(_pre: Self::Pre, _info: DispatchInfo, _len: usize) { ... }
}

Means by which a transaction may be extended. This type embodies both the data and the logic that should be additionally associated with the transaction. It should be plain old data.

Associated Types

type AccountId

The type which encodes the sender identity.

type Call

The type which encodes the call to be dispatched.

type AdditionalSigned: Encode

Any additional data that will go into the signed payload. This may be created dynamically from the transaction using the additional_signed function.

type Pre: Default

The type that encodes information that can be passed from pre_dispatch to post-dispatch.

Loading content...
Required methods

fn additional_signed(&self) -> Result<Self::AdditionalSigned, &'static str>

Construct any additional data that should be in the signed payload of the transaction. Can also perform any pre-signature-verification checks and return an error if needed.

Loading content...
Provided methods

fn validate(
&self,
_who: &Self::AccountId,
_call: &Self::Call,
_info: DispatchInfo,
_len: usize
) -> Result<ValidTransaction, DispatchError>

Validate a signed transaction for the transaction queue.

This function can be called frequently by the transaction queue, to obtain transaction validity against current state. It should perform all checks that determine a valid transaction, that can pay for it's execution and quickly eliminate ones that are stale or incorrect.

Make sure to perform the same checks in pre_dispatch function.

fn pre_dispatch(
self,
who: &Self::AccountId,
call: &Self::Call,
info: DispatchInfo,
len: usize
) -> Result<Self::Pre, DispatchError>

Do any pre-flight stuff for a signed transaction.

Note this function by default delegates to validate, so that all checks performed for the transaction queue are also performed during the dispatch phase (applying the extrinsic).

If you ever override this function, you need to make sure to always perform the same validation as in validate.

fn validate_unsigned(
_call: &Self::Call,
_info: DispatchInfo,
_len: usize
) -> Result<ValidTransaction, DispatchError>

Validate an unsigned transaction for the transaction queue.

Normally the default implementation is fine since ValidateUnsigned is a better way of recognising and validating unsigned transactions.

This function can be called frequently by the transaction queue, to obtain transaction validity against current state. It should perform all checks that determine a valid unsigned transaction, and quickly eliminate ones that are stale or incorrect.

Make sure to perform the same checks in pre_dispatch_unsigned function.

fn pre_dispatch_unsigned(
call: &Self::Call,
info: DispatchInfo,
len: usize
) -> Result<Self::Pre, DispatchError>

Do any pre-flight stuff for a unsigned transaction.

Note this function by default delegates to validate_unsigned, so that all checks performed for the transaction queue are also performed during the dispatch phase (applying the extrinsic).

If you ever override this function, you need to make sure to always perform the same validation as in validate_unsigned.

fn post_dispatch(_pre: Self::Pre, _info: DispatchInfo, _len: usize)

Do any post-flight stuff for a transaction.

Loading content...