Module sr_primitives::weights

Primitives for transaction weighting.

Each dispatch function within decl_module! can have an optional #[weight = $x] attribute. $x can be any type that implements the ClassifyDispatch<T> and WeighData<T> traits. By default, All transactions are annotated with #[weight = SimpleDispatchInfo::default()].

Note that the decl_module macro cannot enforce this and will simply fail if an invalid struct (something that does not implement Weighable) is passed in.

Re-exports

pub use crate::transaction_validity::TransactionPriority;

Structs

DispatchInfo
A bundle of static information collected from the #[weight = $x] attributes.

Enums

DispatchClass
A generalized group of dispatch types. This is only distinguishing normal, user-triggered transactions (Normal) and anything beyond which serves a higher purpose to the system (Operational).

SimpleDispatchInfo
Default type used with the #[weight = x] attribute in a substrate chain.

Traits

ClassifyDispatch
Means of classifying a dispatchable function.

GetDispatchInfo
A Dispatchable function (aka transaction) that can carry some static information along with it, using the #[weight] attribute.

WeighData
Means of weighing some particular kind of data (T).

Type Definitions

Weight
Numeric range of a transaction weight.