Trait sr_primitives::traits::Applyable

pub trait Applyable: Sized + Send + Sync {
    type AccountId: Member + MaybeDisplay;
    type Index: Member + MaybeDisplay + SimpleArithmetic;
    type Call: Member;
    fn index(&self) -> Option<&Self::Index>;

    fn sender(&self) -> Option<&Self::AccountId>;

    fn deconstruct(self) -> (Self::Call, Option<Self::AccountId>);
}

An "executable" piece of information, used by the standard Substrate Executive in order to enact a piece of extrinsic information by marshalling and dispatching to a named function call.

Also provides information on to whom this information is attributable and an index that allows each piece of attributable information to be disambiguated.

Associated Types

type AccountId: Member + MaybeDisplay

Id of the account that is responsible for this piece of information (sender).

type Index: Member + MaybeDisplay + SimpleArithmetic

Index allowing to disambiguate other Applyables from the same AccountId.

type Call: Member

Function call.

Loading content...
Required methods

fn index(&self) -> Option<&Self::Index>

Returns a reference to the index if any.

fn sender(&self) -> Option<&Self::AccountId>

Returns a reference to the sender if any.

fn deconstruct(self) -> (Self::Call, Option<Self::AccountId>)

Deconstructs into function call and sender.

Loading content...